Portable JPA does not allow for inner classes
